+++ docs/src/CHANGELOG.md
@@ -12,6 +12,7 @@
 #### Bug fixes
 - fix string subscripting bug
 - fix empty list parsing bug
+- fix random library bugs
 
 #### Removed